backend/requirements.txt
asgiref==3.3.1; python_version >= '3.5'
cffi==1.14.4; python_version >= '3.5'
cryptography==3.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
django-cors-headers==3.1.1
django-rest-knox==4.1.0
django==3.0
djangorestframework==3.10.3
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.4
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
PyMySQL==1.0.3

